refactor(config): extract config into separate modules
